chore: extract common storybook configuration
