+++ .storybook/preview.jsx
@@ -1,0 +1,37 @@
+import * as React from 'react';
+import { Title, Subtitle, Description, Primary, ArgsTable, PRIMARY_STORY } from '@storybook/addon-docs/blocks';
+import customTheme from './theme';
+
+import './google-fonts.css';
+
+export const globalTypes = {
+  theme: {
+    name: 'Theme',
+    description: 'Global theme for components',
+    defaultValue: 'light',
+    toolbar: {
+      icon: 'circlehollow',
+      items: ['light', 'dark'],
+    },
+  },
+};
+
+export const parameters = {
+  docs: {
+    page: () => (
+      <>
+        <Title />
+        <Subtitle />
+        <Description />
+        <Primary />
+        <ArgsTable story={PRIMARY_STORY} />
+      </>
+    ),
+    theme: customTheme,
+  },
+  options: {
+    storySort: {
+      order: ['Public', 'Internal'],
+    },
+  },
+};
