+++ packages/elements-core/.storybook/manager.js
@@ -1,9 +1,1 @@
-
-
-import { addons } from '@storybook/addons';
-import customTheme from './theme';
-
-addons.setConfig({
-  theme: customTheme,
-  panelPosition: 'right',
-});
+import "../../../.storybook/manager";
