--- packages/elements-core/.storybook/preview-head.html
@@ -1,3 +1,0 @@
-<!-- .storybook/preview-head.html -->
-
-<link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600&amp;display=swap" rel="stylesheet">
