+++ packages/elements-core/.storybook/preview.jsx
@@ -1,24 +1,11 @@
-import * as React from 'react';
-import { Provider as MosaicProvider, subscribeTheme } from '@stoplight/mosaic';
-import { Title, Subtitle, Description, Primary, ArgsTable, PRIMARY_STORY } from '@storybook/addon-docs/blocks';
-import customTheme from './theme';
+export * from '../../../.storybook/preview';
 
+import * as React from "react";
+import {subscribeTheme, Provider as MosaicProvider} from "@stoplight/mosaic";
+import {PersistenceContextProvider, Styled} from "../src";
 import '../src/styles.css';
 
-import { PersistenceContextProvider } from '../src/context/Persistence';
-import { Styled } from '../src/styled';
 
-export const globalTypes = {
-  theme: {
-    name: 'Theme',
-    description: 'Global theme for components',
-    defaultValue: 'light',
-    toolbar: {
-      icon: 'circlehollow',
-      items: ['light', 'dark'],
-    },
-  },
-};
 
 const ThemeProvider = (Story, context) => {
   const theme = context.globals.theme;
@@ -30,12 +17,12 @@ const ThemeProvider = (Story, context) => {
   );
 };
 
-const PersistenceBoundaryDecorator = (Story) => (
-  <PersistenceContextProvider><Story /></PersistenceContextProvider>
-);
-
 const MosaicProviderDecorator = (Story) => (
   <MosaicProvider><Story/></MosaicProvider>
+);
+
+const PersistenceBoundaryDecorator = (Story) => (
+  <PersistenceContextProvider><Story /></PersistenceContextProvider>
 );
 
 const StyledDecorator = (Story) => (
@@ -43,23 +30,3 @@ const StyledDecorator = (Story) => (
 );
 
 export const decorators = [ThemeProvider, MosaicProviderDecorator, PersistenceBoundaryDecorator, StyledDecorator];
-
-export const parameters = {
-  docs: {
-    page: () => (
-      <>
-        <Title />
-        <Subtitle />
-        <Description />
-        <Primary />
-        <ArgsTable story={PRIMARY_STORY} />
-      </>
-    ),
-    theme: customTheme,
-  },
-  options: {
-    storySort: {
-      order: ['Public', 'Internal'],
-    },
-  },
-};
