+++ packages/elements-core/package.json
@@ -26,7 +26,7 @@
     "postbuild": "yarn build.styles",
     "build.react": "sl-scripts build",
     "build.styles": "node ../../scripts/build-styles.js",
-    "build.docs": "build-storybook -c .storybook -o dist-storybook",
+    "build.docs": "build-storybook -o dist-storybook",
     "commit": "git-cz",
     "release": "sl-scripts release",
     "release.docs": "sl-scripts release:docs",
